15th commit : - fixed FlowRemoteMediatorFragment loading first two pages only

+++ app/src/main/java/com/eslamhamdi/paging3_todo/Utils.kt
@@ -32,7 +32,7 @@ fun List<DataItem?>?.toEntities(): List<TaskEntity>?
     return this?.map {
 
         TaskEntity(note = it?.note,updatedAt = it?.updatedAt,userId = it?.userId,createdAt = it?.createdAt
-            ,id = it?.id?.toLong(),title = it?.title,body = it?.body,status = it?.status)
+            ,id = it?.id?.toInt(),title = it?.title,body = it?.body,status = it?.status)
     }
 }
